Rewrote add_samples() so 8, 16 and 32 bit samples can be used

+++ scripts/wave.h
@@ -28,7 +28,7 @@
     } Wave;
 
     // Method to return a Wave instance that is ready to be written to a wav file. For simplicity it has only one channel.
-    Wave make_wave(const int16_t* sample_array, uint32_t numberof_samples, uint16_t numberof_channels, uint32_t sample_rate, uint16_t bits_per_sample);
+    Wave make_wave(const int32_t* sample_array, uint32_t numberof_samples, uint16_t numberof_channels, uint32_t sample_rate, uint16_t bits_per_sample);
 
     // Method to deallocate the memory used to hold the sample in data while the file is written to the wav file.
     void remove_sample_data(Wave* wave);
